Refactor styling and first endpoints working
