+++ src/client/components/champions/ChampionSelector/SelectedTeamGroup.tsx
@@ -1,7 +1,7 @@
 import { Flex } from '@mantine/core';
-import SelectedChampionCard from '@/client/components/champions/SelectedChampionCard';
-import { useSelectedTeam } from '../../contexts/SelectedTeamContext';
-import styles from '@/client/css/styles.module.css';
+import SelectedChampionCard from './SelectedChampionCard';
+import { useSelectedTeam } from '../../../contexts/SelectedTeamContext';
+import styles from './ChampionSelector.module.css';
 
 interface SelectedTeamGroupProps {
   imageWidth: number;
